CM-111 and general fixes
