+++ src/store/modules/establishment.module.js
@@ -1,5 +1,7 @@
 import EstablishmentService from '../../services/Administration/establishment.service';
 import UsersService from '../../services/users/user.service';
+import {getHttpError} from "../../services/http-common";
+import {ADMINISTRATION_ERRORS} from "../../constants";
 
 const state = {
     establishment: {name: null, cuig: null, location: null, phone: null, email: null, ownerUserName: "" },
@@ -12,34 +14,7 @@ const mutations = {
         state.error = {type: null, message: null};
     },
     setError(state, error) {
-        let message = 'ERROR indefinido, intente nuevamente mas tarde.';
-        let payload;
-        if(error.response){
-            if([404, 401, 409].includes(error.response.status)){
-                payload = error.response.statusText;
-            }
-        }else{
-            payload = error.message;
-        }
-        console.error(error);
-        switch (payload){
-            case null:
-                return state.error = {type: null, message: null};
-            case "Unauthorized":
-                message = 'No esta autorizado';
-                break;
-            case "Conflict":
-                message = 'El establecimiento que esta intentando registrar ya existe';
-                break;
-            case "Not Found":
-                message = 'Establecimiento no encontrado';
-                break;
-            case "Server":
-                message = 'ERROR en el servidor, intente nuevamente mas tarde.';
-                break;
-
-        }
-        return state.error = {type: 'Server', message: message}
+        return state.error = getHttpError(ADMINISTRATION_ERRORS, error.response.status);
     }
 }
